examples/imagenet_pruning/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pypi_0
backcall=0.2.0=pypi_0
betty-ml=0.2.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=39.0.1=py39h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
executing=1.2.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
flit-core=3.8.0=py39h06a4308_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
h5py=3.7.0=py39h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
htop=3.2.2=h8228510_0
idna=3.4=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.13=pypi_0
ipython=8.12.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=py39h06a4308_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnl=3.7.0=h166bdaf_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_1
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.22.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.0=pypi_0
pyopenssl=23.0.0=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_2
pytorch=2.0.0=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_3
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_1
sentry-sdk=1.19.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.41.1=h5eee18b_0
stack-data=0.6.2=pypi_0
sympy=1.11.1=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torchaudio=2.0.0=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu117
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py39h06a4308_0
wandb=0.14.2=pypi_0
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.4=hc292b87_0
